requirements.txt
tk
pysubs2
opencc-python-reimplemented
ffmpeg-python
pymkv
argostranslate
pySide6
langdetect